requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
django-simple-history==3.0.0
djangorestframework==3.12.4
idna==2.10
mailchimp-transactional==1.0.26
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
